Make tests faster by splitting out CI tests into multiple files
CI test is the slowest test, therefore the whole suite is dependent on the amount of time it takes to run.

Previously we were doing 3 CI test deployments in one file, we can make the suite much faster by instead splitting them into 3 different files.

I think we could further speed things up by maybe turning off the retry behavior of the CI test that we know will fail?
