+++ .travis.yml
@@ -3,7 +3,7 @@ rvm:
 - 2.3.1
 before_script:
   - bundle exec hatchet ci:setup
-script: bundle exec parallel_test test/hatchet -n 9
+script: bundle exec parallel_test test/hatchet -n 11
 after_script: bundle exec rake hatchet:teardown_travis
 env:
   global:
