+++ hatchet.gemspec
@@ -29,6 +29,6 @@ Gem::Specification.new do |gem|
   gem.add_development_dependency "minitest",       "~> 5.1"
   gem.add_development_dependency "rake",           "~> 10"
   gem.add_development_dependency "mocha",          "~> 1"
-  gem.add_development_dependency "parallel_tests", "~> 0"
+  gem.add_development_dependency "parallel_tests", "~> 2"
   gem.add_development_dependency "travis",         "~> 1"
 end
