+++ test/test_helper.rb
@@ -23,3 +23,5 @@ end
 
 
 ENV['HATCHET_BUILDPACK_BRANCH'] = "master"
+
+require 'parallel_tests/test/runtime_logger' if ENV['RECORD_RUNTIME']
