+++ tmp/parallel_runtime_test.log
@@ -1,0 +1,3 @@
+test/hatchet/edit_repo_test.rb:5.40
+test/hatchet/local_repo_test.rb:9.60
+test/hatchet/heroku_api_test.rb:21.88
